gamewithboard
Write a class called Game. It should have the following data:

a 2D integer array called board
an integer called pos_row to store the row of the player’s position
an integer called pos_col to store the column of the player’s position
Also, it should have a constructor that takes as parameter 3 integers

First, the size of the board
Second, the row of the player’s starting position
Third, the column of the player’s starting position
Finally, you should also define a method called draw: it prints the board but does not return anything.

As per the parameters, you have to draw the board. The board is full of zeros, except for the position of the player. The position of the player is marked by 1.

Only write the Game class, assume the Main class has already been written.